# Add detailed backtest performance results in CLI