--- a/src/Domain/Backtesting/Event/BacktestPhaseEvent.php
+++ b/src/Domain/Backtesting/Event/BacktestPhaseEvent.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Stochastix\Domain\Backtesting\Event;
+
+namespace Stochastix\Domain\Backtesting\Event;
+
+use Symfony\Contracts\EventDispatcher\Event;
+
+final class BacktestPhaseEvent extends Event
+{
+    public function __construct(
+        public readonly string $runId,
+        public readonly string $phase,
+        public readonly string $eventType,
+    ) {
+    }
+}