This PR enhances the `stochastix:backtest` command output with detailed performance results:

```
Execution Profile
-----------------

 ---------------- ----------- ----------
  Phase            Duration    Memory
 ---------------- ----------- ----------
  Configuration    0.30 ms     44.50 MB
  Initialization   33.20 ms    56.50 MB
  Loop             304.10 ms   80.50 MB
  Statistics       971.10 ms   90.50 MB
  Saving           303.90 ms   90.50 MB
 ---------------- ----------- ----------

📊 Backtest finished in: 1612.60 ms / Peak Memory usage: 90.50 MB
```